updated code v2
