+++ interfaces/lPushSupport.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Interfaces;
+
+interface lPushSupport
+{
+    public function lpush(string $key, string $value);
+}
+
